+++ backend/hqlib/requirement/product_quality.py
@@ -50,12 +50,10 @@ class TrustedProductMaintainability(Requirement):
 
 
 class UnitTests(Requirement):
-    """ Require the product to have unit and/or integration tests. """
-    _name = 'Unit and/or integration tests'
+    """ Require the product to have unit tests. """
+    _name = 'Unit tests'
     _metric_classes = (metric.FailingUnittests, metric.UnittestReportAge,
-                       metric.UnittestLineCoverage, metric.UnittestBranchCoverage,
-                       metric.IntegrationtestLineCoverage, metric.IntegrationtestBranchCoverage,
-                       metric.UnitAndIntegrationTestLineCoverage, metric.UnitAndIntegrationTestBranchCoverage)
+                       metric.UnittestLineCoverage, metric.UnittestBranchCoverage)
 
 
 class ART(Requirement):